fix(preload): do not copy preload.js to tmp user data dir
It is not possible to use path with spaces as -r argument, so we
used to copy this file to user data dir that was located in place
without spaces.

Alternative solution: put preload.js in separate folder and add
this folder to NODE_PATH, use -r ndb/preload.js afterwards.